+++ Pandora/Paradigm/Controlflow/Joint/Schemes/TUT.hs
@@ -3,8 +3,8 @@ module Pandora.Paradigm.Controlflow.Joint.Schemes.TUT (TUT (..)) where
 import Pandora.Core.Functor (type (:.), type (:=))
 import Pandora.Paradigm.Controlflow.Joint.Interpreted (Interpreted (Primary, run))
 
-newtype TUT ct cu cv t u t' a = TUT (t :. u :. t' := a)
+newtype TUT ct ct' cu t t' u a = TUT (t :. u :. t' := a)
 
-instance Interpreted (TUT ct cu cv t u t') where
-	type Primary (TUT ct cu cv t u t') a = t :. u :. t' := a
+instance Interpreted (TUT ct ct' cu t t' u) where
+	type Primary (TUT ct ct' cu t t' u) a = t :. u :. t' := a
 	run (TUT x) = x