+++ Pandora/Paradigm/Structure.hs
@@ -1,6 +1,6 @@
 module Pandora.Paradigm.Structure (module Exports) where
 
-import Pandora.Paradigm.Structure.Nonempty as Exports
+import Pandora.Paradigm.Structure.Variation.Nonempty as Exports
 import Pandora.Paradigm.Structure.Specific.Binary as Exports
 import Pandora.Paradigm.Structure.Specific.Graph as Exports
 import Pandora.Paradigm.Structure.Specific.Stack as Exports