+++ Pandora/Paradigm/Structure/Specific.hs
@@ -1,6 +1,0 @@
-module Pandora.Paradigm.Structure.Specific (module Exports) where
-
-import Pandora.Paradigm.Structure.Specific.Stream as Exports
-import Pandora.Paradigm.Structure.Specific.Stack as Exports
-import Pandora.Paradigm.Structure.Specific.Graph as Exports
-import Pandora.Paradigm.Structure.Specific.Binary as Exports