+++ Pandora/Paradigm/Structure/Stream.hs
@@ -1,4 +1,4 @@
-module Pandora.Paradigm.Structure.Specific.Stream (Stream) where
+module Pandora.Paradigm.Structure.Stream (Stream) where
 
 import Pandora.Paradigm.Basis.Identity (Identity)
 import Pandora.Paradigm.Basis.Twister (Twister)