+++ Pandora/Paradigm/Structure/Variation/Nonempty.hs
@@ -4,9 +4,9 @@ import Pandora.Paradigm.Basis.Maybe (Maybe)
 import Pandora.Paradigm.Basis.Edges (Edges)
 import Pandora.Paradigm.Basis.Twister (Twister)
 import Pandora.Paradigm.Basis.Wye (Wye)
-import Pandora.Paradigm.Structure.Specific.Binary (Binary)
-import Pandora.Paradigm.Structure.Specific.Graph (Graph)
-import Pandora.Paradigm.Structure.Specific.Stack (Stack)
+import Pandora.Paradigm.Structure.Binary (Binary)
+import Pandora.Paradigm.Structure.Graph (Graph)
+import Pandora.Paradigm.Structure.Stack (Stack)
 
 -- | Type synonymous for at least one element data structure
 type family Nonempty (structure :: * -> *) where