+++ Pandora/Paradigm/Basis/Kan.hs
@@ -1,25 +1,28 @@
-module Pandora.Paradigm.Basis.Kan (Lan (..), Ran (..)) where
+module Pandora.Paradigm.Basis.Kan (Kan (..)) where
 
 import Pandora.Pattern.Category ((.))
 import Pandora.Pattern.Functor.Contravariant (Contravariant ((>$<)))
 import Pandora.Pattern.Functor.Covariant (Covariant ((<$>)))
 import Pandora.Pattern.Functor.Divariant (($))
 import Pandora.Paradigm.Controlflow.Joint.Interpreted (Interpreted (Primary, run))
+import Pandora.Paradigm.Basis.Wye (Wye (Left, Right))
 
-newtype Lan t u b a = Lan ((t b -> a) -> u b)
+data family Kan (v :: * -> k) (t :: * -> *) (u :: * -> *) b a
 
-instance Contravariant (Lan t u b) where
+data instance Kan 'Left t u b a = Lan ((t b -> a) -> u b)
+
+instance Contravariant (Kan 'Left t u b) where
 	f >$< Lan x = Lan $ x . (f .)
 
-instance Interpreted (Lan t u b) where
-	type Primary (Lan t u b) a = (t b -> a) -> u b
+instance Interpreted (Kan 'Left t u b) where
+	type Primary (Kan 'Left t u b) a = (t b -> a) -> u b
 	run (Lan x) = x
 
-newtype Ran t u b a = Ran ((a -> t b) -> u b)
+data instance Kan 'Right t u b a = Ran ((a -> t b) -> u b)
 
-instance Covariant (Ran t u b) where
+instance Covariant (Kan 'Right t u b) where
 	f <$> Ran x = Ran $ x . (. f)
 
-instance Interpreted (Ran t u b) where
-	type Primary (Ran t u b) a = (a -> t b) -> u b
+instance Interpreted (Kan 'Right t u b) where
+	type Primary (Kan 'Right t u b) a = (a -> t b) -> u b
 	run (Ran x) = x