+++ CHANGELOG.md
@@ -188,3 +188,4 @@
 
 # 0.2.6
 * Change order of arguments for `TUT` joint schema
+* Binary trees can be empty