+++ CHANGELOG.md
@@ -188,4 +188,5 @@
 
 # 0.2.6
 * Change order of arguments for `TUT` joint schema
-* Binary trees can be empty
+* `Binary` trees can be empty
+* Define `left_sub_tree` and `right_sub_tree` lenses of `Binary` tree