+++ CHANGELOG.md
@@ -196,3 +196,4 @@
 * Remove `Specific` module
 * Add functional dependency to `Nonempty` type family and distribute instances
 * Define `left` and `right` natural transformations from `Wye` to `Maybe`
+* Define experimental `Substructure` type class