+++ CHANGELOG.md
@@ -199,3 +199,4 @@
 * Define experimental `Substructure` type class
 * Remove `left_sub_tree` and `right_sub_tree` lenses of `Binary` tree
 * Define `this` and `that` methods from `Variation` to `Maybe`
+* Define `Backwards` type to run `Applicative` actions in reverse order